divide file into small pieces
